Fixed stuff to cross compile for windows.

+++ bitmap.cpp
@@ -132,9 +132,9 @@ static void handle_bitmap_header(const RECT *r, const BITMAPFILEHEADER *bf, cons
 }
 
 #ifdef _WIN32
-static unsigned int log2(unsigned int value)
+static WORD log2(WORD value)
 {
-	unsigned int l = 0;
+	WORD l = 0;
 	while( (value >> l) > 1 ) ++l;
 	return l;
 }
